+++ octoprint_PrusaMeshMap/static/js/PrusaMeshMap.js
@@ -14,7 +14,7 @@ $(function() {
         self.controlViewModel = parameters[2];
 
 	// Assign matplotlib color dropdown options
-	self.matplotlib_heatmap_theme_options = ["viridis", "plasma", "inferno", "magma"];
+	self.matplotlib_heightmap_theme_options = ["viridis", "plasma", "inferno", "magma"];
 
         // TODO: Implement your plugin's view model here.
         self.sendPrusaBedLevel = function() {